# swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
paths:
...
  /receipt/:
    post:
      tags:
        - Receipt
      summary: Create receipt
      description: This will create a receipt [SYSTEM USER]
      requestBody:
        description: Receipt to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertReceiptCommand"
      operationId: createReceipt
      responses:
        200:
          description: The created receipts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
...
  /receipt/{receiptId}:
    parameters:
      - in: path
        name: receiptId
        schema:
          type: integer
        required: true
        description: Id of receipt to get
    get:
      tags:
        - Receipt
      summary: Get receipt
      description: This will get a receipt by receipt id [SYSTEM USER]
      operationId: getReceiptById
      responses:
        200:
          description: The receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - Receipt
      summary: Update receipt
      description: This will update a receipt by receipt id [SYSTEM USER]
      requestBody:
        description: Receipt to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertReceiptCommand"
      operationId: updateReceipt
      responses:
        200:
          description: The updated receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Receipt
      summary: Delete receipt
      description: This will delete a receipt by id [SYSTEM USER]
      operationId: deleteReceiptById
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /receipt/{receiptId}/duplicate:
    post:
      tags:
        - Receipt
      summary: Duplicate receipt
      description: This will duplicate a receipt [SYSTEM USER]
      parameters:
        - in: path
          name: receiptId
          schema:
            type: integer
          required: true
          description: Id of receipt to duplicate
      operationId: duplicateReceipt
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /receipt/bulkStatusUpdate:
    post:
      tags:
        - Receipt
      summary: Bulk receipt status update
      description:
        This will bulk update receipt statuses with the option of adding
        a comment to each [SYSTEM USER]
      operationId: bulkReceiptStatusUpdate
      requestBody:
        description: Bulk status data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkStatusUpdateCommand"
      responses:
        200:
          description: The updated receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Request was successfully processed
...
    Internal:
      description: There was an error processing the request
  schemas:
    ReceiptStatus:
      description: Status of a receipt
      type: string
      enum:
        - "OPEN"
        - "NEEDS_ATTENTION"
        - "RESOLVED"
        - "DRAFT"
        - ""
    ItemStatus:
      type: string
      enum:
        - "OPEN"
        - "RESOLVED"
        - "DRAFT"
...
    Category:
      type: object
      properties:
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Name of the category
        description:
          type: string
          description: Description of the category
        updatedAt:
          type: string
          x-go-name: UpdatedAt
      description: Category to relate receipts to
    UpsertCategoryCommand:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          description: Category id
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
...
    Comment:
      required:
        - comment
        - receiptId
        - userId
        - id
      type: object
      properties:
        additionalInfo:
          type: string
          description: Additional information about the comment
        comment:
          type: string
          description: Comment itself
        commentId:
          type: integer
          description: Comment foreign key used for repleis
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        receiptId:
          type: integer
          description: Receipt foreign key
        updatedAt:
          type: string
        userId:
          type: integer
          description: User foreign key
          format: uint64
      description: User comment left on receipts
    FileData:
      required:
        - receiptId
        - id
      type: object
      properties:
        createdAt:
          type: string

          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
          x-go-name: CreatedBy
        fileType:
          type: string
          description: MIME file type
          x-go-name: FileType
        id:
          type: integer
          format: uint64
          x-go-name: ID
        imageData:
          type: array
          description: Image data
          items:
            type: integer
            format: uint8
          x-go-name: ImageData
        name:
          type: string
          description: File name
          x-go-name: Name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
          x-go-name: ReceiptId
        size:
          type: integer
          description: File size
          format: uint64
          x-go-name: Size
        updatedAt:
          type: string

          x-go-name: UpdatedAt
      description: File data for images on a receipt
      x-go-package: receipt-wrangler/api/internal/models
...
    Item:
      required:
        - amount
        - chargedToUserId
        - name
        - receiptId
        - status
      type: object
      properties:
        IsTaxed:
          type: boolean
          description: Is taxed (not used)
        amount:
          type: string
          description: Amount the item costs
        chargedToUserId:
          type: integer
          description: User foreign key
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Item name
        receiptId:
          type: integer
          description: Receipt foreign key
          format: uint64
        status:
          type: string
          $ref: "#/components/schemas/ItemStatus"
        categories:
          type: array
          description: Categories associated to the item
          items:
            $ref: "#/components/schemas/Category"
        tags:
          type: array
          description: Tags associated to the item
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
      description: Itemized item on a receipt
      x-go-package: receipt-wrangler/api/internal/models
...
    Receipt:
      required:
        - amount
        - date
        - groupId
        - name
        - paidByUserId
        - status
        - id
      type: object
      properties:
        amount:
          type: string
          description: Receipt total amount
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/Category"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
          format: uint64
        id:
          type: integer
          format: uint64
        imageFiles:
          type: array
          description: Files associated to receipt
          items:
            $ref: "#/components/schemas/FileData"
        name:
          type: string
          description: Receipt name
        paidByUserId:
          type: integer
          description: User paid foreign key
          format: uint64
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/Item"
        resolvedDate:
          type: string
          description: Date resolved
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/Tag"
        updatedAt:
          type: string
        createdByString:
          type: string
          description: Created by string, which is anything that is not a user
      description: Receipt
    Tag:
      required:
        - name
      type: object
      properties:
        createdAt:
          type: string
          x-go-name: CreatedAt
        createdBy:
          type: integer
          format: uint64
        id:
          type: integer
          format: uint64
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
        updatedAt:
          type: string
      description: Tag to relate receipts to
...
    UpsertTagCommand:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          description: Tag id
        name:
          type: string
          description: Tag name
        description:
          type: string
          description: Tag description
      description: Tag to relate receipts to
...
    BulkStatusUpdateCommand:
      required:
        - status
        - receiptIds
      type: object
      properties:
        comment:
          type: string
          description: Optional comment to leave on each receipt
        status:
          type: string
          description: Status to update to
        receiptIds:
          type: array
          items:
            type: integer
          description: Receipt ids to update
...
    UpsertReceiptCommand:
      type: object
      required:
        - name
        - amount
        - date
        - groupId
        - paidByUserId
        - status
      properties:
        name:
          type: string
          description: Receipt name
        amount:
          type: string
          description: Receipt total amount
        date:
          type: string
          description: Receipt date
        groupId:
          type: integer
          description: Group foreign key
        paidByUserId:
          type: integer
          description: User paid foreign key
        status:
          $ref: "#/components/schemas/ReceiptStatus"
        categories:
          type: array
          description: Categories associated to receipt
          items:
            $ref: "#/components/schemas/UpsertCategoryCommand"
        tags:
          type: array
          description: Tags associated to receipt
          items:
            $ref: "#/components/schemas/UpsertTagCommand"
        receiptItems:
          type: array
          description: Items associated to receipt
          items:
            $ref: "#/components/schemas/UpsertItemCommand"
        comments:
          type: array
          description: Comments associated to receipt
          items:
            $ref: "#/components/schemas/UpsertCommentCommand"
    UpsertItemCommand:
      type: object
      required:
        - amount
        - chargedToUserId
        - name
        - receiptId
        - status
      properties:
        amount:
          type: string
          description: Amount the item costs
        chargedToUserId:
          type: integer
          description: User foreign key
        name:
          type: string
          description: Item name
        receiptId:
          type: integer
          description: Receipt foreign key
        status:
          $ref: "#/components/schemas/ItemStatus"
        categories:
          type: array
          description: Categories associated to item
          items:
            $ref: "#/components/schemas/UpsertCategoryCommand"
        tags:
          type: array
          description: Tags associated to item
          items:
            $ref: "#/components/schemas/UpsertTagCommand"
    UpsertCommentCommand:
      type: object
      required:
        - comment
        - receiptId
      properties:
        comment:
          type: string
          description: Comment itself
        receiptId:
          type: integer
          description: Receipt foreign key
        userId:
          type: integer
          description: User foreign key
...